🐞 fix: 重构一些优化过度及优化错误
